Explain with buffering and indenting improvements.
