+++ lib/index.js
@@ -1,9 +1,9 @@
 /**
  * Module dependencies.
  */
-var compiler = require('./compiler'),
-	fs = require('fs'),
+var fs = require('fs'),
 	inProduction = process.env.NODE_ENV === 'production',
+	parser = require('./parser'),
 	path = require('path'),
 	runtime = require('./runtime'),
 	templates = {};
@@ -56,21 +56,30 @@ module.exports = function (path, options, cb) {
  * Compile a template.
  *
  * @param template The template.
+ * @param compress Indicates whether compression has been enabled.
  * @return The compiled template.
  */
-module.exports.compile = function (template) {
-	// Compile a template.
-	var compiled = new Function('runtime', 'root', 'inputPartials', compiler(template, module.exports.options.compress));
-	// Validate and set the runtime render function.
-	runtime.render = runtime.render ? runtime.render : function (root, inputPartials, file) {
-		// Compile, render and return a template from file.
-		return module.exports.compileFile(file)(root, inputPartials);
-	};
-	// Return the scope.
-	return function (root, inputPartials) {
-		// Execute the template.
-		return compiled(runtime, root, inputPartials);
-	};
+module.exports.compile = function (template, compress) {
+	// Validate and set compression.
+	compress = compress === undefined ? module.exports.options.compress : compress;
+	// CHeck if the render function of runtime is invalid.
+	if (!runtime.render) {
+		// Set the runtime render function.
+		runtime.render = function (root, inputPartials, file) {
+			// Compile, render and return a template from file.
+			return module.exports.compileFile(file)(root, inputPartials);
+		};
+	}
+	// Return the invoked parser scope.
+	return (function () {
+		// Parse a template and compile it.
+		var compiled = new Function('runtime', 'root', 'inputPartials', parser(template, module.exports.options.compress));
+		// Return the wrapped function.
+		return function (root, inputPartials) {
+			// Execute the template.
+			return compiled(runtime, root, inputPartials);
+		};
+	})();
 };
 
 /**
@@ -115,6 +124,52 @@ module.exports.compileFile = function (file, directory, cache) {
 	// Throw an error when the file has not been found.
 	throw Error('Invalid template file "' + file + '"');
 };
+
+/**
+ * Explain a template using the compiled function.
+ *
+ * @param template The template.
+ * @param cb The optional callback, or console.log.
+ */
+module.exports.explain = function (template, cb) {
+	// Validate and set the callback
+	cb = cb === undefined ? console.log : cb;
+	// Parse a template and prepare it with a function wrapper.
+	cb('function (runtime, root, inputPartials) {' + '\n' + parser(template, true) + '\n' + '}');
+}
+
+/**
+ * Explain a template from file using the compiled function.
+ *
+ * @param file The file.
+ * @param directory The optional directory.
+ * @param cb The optional callback, or console.log.
+ */
+module.exports.explainFile = function (file, directory, cb) {
+	// Check if the directory is a function.
+	if (typeof directory === 'function') {
+		// Set the callback.
+		cb = directory;
+		// Remove the directory.
+		directory = undefined;
+	}
+	// Validate and set the callback
+	cb = cb === undefined ? console.log : cb;
+	// Validate and set the directory.
+	directory = directory === undefined ? module.exports.options.directory : (Array.isArray(directory) ? directory : [directory]);
+	// Validate and set the file name.
+	file = path.extname(file).length ? file : file + '.' + module.exports.options.extension;
+	// Iterate through each directory.
+	for (var i = 0; i < directory.length; i += 1) {
+		// Check if the template exists.
+		if (fs.existsSync(path.join(directory[i], file))) {
+			// Parse a template and prepare it with a function wrapper.
+			cb('function (runtime, root, inputPartials) {' + '\n' + parser(fs.readFileSync(path.join(directory[i], file), 'utf8'), true) + '\n' + '}');
+			// Return.
+			return;
+		}
+	}
+}
 
 /**
  * The options.
