+++ lib/parser.js
@@ -1,230 +1,224 @@
 /**
- * Represents the module performing parsing. The parser is used by the compiler to parse the
- * piece of information with a status assigned by lexical anylsis and fills the state with parsed
- * instructions the compiler can use to create a function.
+ * Represents the module performing compilation. The compiler makes use of lexical
+ * analysis to understand the template structure and extracts template pieces using
+ * that understanding. Each template piece is categorized and forwarded to the emitter,
+ * which emits functional code.
  */
-var utils = require('./utils')
+var lexer = require('./lexer'),
+	emitter = require('./emitter'),
+	State = require('./state');
 
 /**
- * The statements.
- */
-var statements = {
-	// Include the include statement.
-	'data-include': require('./statements/include-statement'),
-	// Include the partial statement.
-	'data-partial': require('./statements/partial-statement'),
-	// Include the if statement.
-	'data-if': require('./statements/if-statement'),
-	// Include the each statement.
-	'data-each': require('./statements/each-statement')
-};
-
-/**
- * Parse a beginning statement.
+ * Compile the template.
  *
- * @param state The state.
- * @param attributes The attributes.
- * @returns Indicates whether the element is allowed to be disposed of.
+ * @param view The template.
+ * @param compress Indicates whether compression has been enabled.
+ * @returns The compiled template.
  */
-module.exports.begin = function (state, attributes) {
-	// Initialize the attribute to delete.
-	var attributesToDelete = [];
-	// Arrange attributes to a ensure a correct order.
-	arrange(attributes);
-	// Set the element insert to invalid.
-	state.insert = -1;
-	// Iterate through each attribute.
-	for (var i = 0; i < attributes.length; i += 1) {
-		// Check if the attribute is a recognized statement.
-		if (attributes[i].key in statements && statements[attributes[i].key] && typeof statements[attributes[i].key].begin === 'function') {
-			// Begin writing the statement.
-			statements[attributes[i].key].begin(state, attributes[i]);
-			// Push the attribute to the attribute deletion list.
-			attributesToDelete.push(attributes[i].key);
-		}
-	}
-	// Iterate through each attribute candidate for deletion.
-	for (var i = 0; i < attributesToDelete.length; i += 1) {
-		// Iterate through each attribute.
-		for (var j = 0; j < attributes.length; j += 1) {
-			// Check if the attribute should be deleted.
-			if (attributes[j].key === attributesToDelete[i]) {
-				// Delete the attribute.
-				attributes.splice(j, 1);
-				// Break from iteration;
-				break;
-			}
-		}
-	}
-	// Return the status indicating whether the element is allowed to be disposed of.
-	return state.insert !== -1;
-};
-
-/**
- * Parse an ending statement.
- *
- * @param state The state.
- * @param attributes The attributes.
- */
-module.exports.end = function (state, attributes) {
-	// Arrange attributes to a ensure a proper emitting.
-	arrange(attributes, true);
-	// Iterate through each attribute.
-	for (var i = 0; i < attributes.length; i += 1) {
-		// Check if the attribute is a recognized statement.
-		if (attributes[i].key in statements && statements[attributes[i].key] && typeof statements[attributes[i].key].end === 'function') {
-			// End writing the statement.
-			statements[attributes[i].key].end(state, attributes[i]);
-		}
-	}
-};
-
-/**
- * Parse the finalized instructions and return the function body.
- *
- * @param state The state.
- * @returns The function body.
- */
-module.exports.finalize = function (state) {
-	// Push the instruction to define the result.
-	state.instructions.unshift('var result = \'\';');
-	// Push the instruction to define the handlers.
-	state.instructions.unshift('var handlers = runtime.handlers;');
-	// Push the instruction to define the filters.
-	state.instructions.unshift('var filters = runtime.filters;');
-	// Push the instruction to define the data.
-	state.instructions.unshift('var data = root;');
-	// Push the instruction to join and return the result.
-	state.instructions.push('return result;');
-	// Returned the instructions.
-	return state.instructions.join('');
-};
-
-/**
- * Parse an invocation statement.
- *
- * @param state The state.
- * @param attributes The attributes.
- */
-module.exports.invoke = function (state, attributes) {
-	// Initialize the result.
-	var result = false;
-	// Arrange attributes to a ensure a proper emitting.
-	arrange(attributes);
-	// Iterate through each attribute.
-	for (var i = 0; i < attributes.length; i += 1) {
-		// Check if the attribute is a recognized statement.
-		if (attributes[i].key in statements && statements[attributes[i].key] && typeof statements[attributes[i].key].invoke === 'function') {
-			// End writing the statement.
-			statements[attributes[i].key].invoke(state, attributes[i]);
-			// Delete the attribute.
-			delete attributes[i];
-			// Set the result.
-			result = true;
-		}
-	}
-	// Return the result.
-	return result;
-};
-
-/**
- * Parse text.
- *
- * @param state The state.
- * @param text The text.
- * @param isOpening Indicates whether the variable is contained in an opening literal.
- */
-module.exports.text = function (state, text, isOpening) {
-	// Initialize the result.
-	var result = 'result += \'' + utils.inline(text) + '\';';
-	// Check if the variable is contained in an opening literal and is to be inserted.
-	if (state.insert !== -1 && isOpening) {
-		// Insert the text.
-		state.instructions.splice(state.insert, 0, result);
-		// Increment the insert location.
-		state.insert += 1;
-	}
-	// Otherwise it is a regular variable.
-	else {
-		// Push the text.
-		state.instructions.push(result);
-	}
-};
-
-/**
- * Parse a variable.
- *
- * @param state The state.
- * @param name The variable.
- * @param handlers Each handler.
- * @param escapeHtml Indicates whether escaping of HTML is required.
- * @param isOpening Indicates whether the variable is contained in an opening literal.
- */
-module.exports.variable = function (state, name, handlers, escapeHtml, isOpening) {
-	// Initialize the result.
-	var result = name;
-	// Check if escaping of HTML is required.
-	if (escapeHtml) {
-		// Push the escape handler to the handlers.
-		handlers.push('escape');
-	}
-	// Iterate through each handler.
-	for (var i = 0; i < handlers.length; i += 1) {
-		// Set the result to include the handler invocation.
-		result = 'handlers.' + handlers[i] + '(' + result + ')';
-	}
-	// Set the result to include checking for an undefined variable.
-	result = name + ' === undefined ? \'\' : ' + result;
-	// Check if the variable is contained in an opening literal and is to be inserted.
-	if (state.insert !== -1 && isOpening) {
-		// Insert the variable.
-		state.instructions.splice(state.insert, 0, 'result += ' + result + ';');
-		// Increment the insert location.
-		state.insert += 1;
-	}
-	// Otherwise it is a regular variable.
-	else {
-		// Push the variable.
-		state.instructions.push('result += ' + result + ';');
-	}
-};
-
-/**
- * Arrange attributes to a ensure a correct order.
- *
- * @param attributes The attributes.
- * @param reverse Indicates whether the sorting is to be reversed.
- */
-function arrange(attributes, reverse) {
-	// Sort the attributes.
-	attributes.sort(function (a, b) {
-		// Retrieve the priority and sort accordingly.
-		return reverse ? find(a.key) < find(b.key) : find(a.key) > find(b.key);
-	});
+module.exports = function (template, compress) {
+	// Validate and clean the template.
+	template = compress ? clean(template) : template;
+	// Compile the template for each association and section.
+	return emitter.finalize(compileTemplate(State(), template, lexer(template), 0, template.length));
 }
 
 /**
- * Find the priority index for the search input.
+ * Clean the template.
  *
- * @param search The search input.
- * @returns The priority index.
+ * @param template The template.
+ * @returns The template.
  */
-function find(search) {
-	// Initialize the index.
-	var index = 0;
-	// Iterate through each statement.
-	for (var key in statements) {
-		// Check if the key is a property of the statements object.
-		if (statements.hasOwnProperty(key)) {
-			// Check if the key matches the search input.
-			if (key === search) {
-				// Return the index.
-				return index;
+function clean(template) {
+	// Remove each comment from the template.
+	template = template.replace(/<!--[^\[]*?-->/gi, '');
+	// Remove remaining duplicate whitespace.
+	template = template.replace(/\s+/gi, ' ');
+	// Remove each whitespace between elements.
+	template = template.replace(/>\s+</gi, '><');
+	// Return the template.
+	return template;
+}
+
+/**
+ * Compile the literal into variables and text.
+ *
+ * @param state The state.
+ * @param literal The literal.
+ * @param isOpeningElement Indicates whether the literal is an opening element.
+ * @returns The state.
+ */
+function compileLiteral(state, literal, isOpeningElement) {
+	// Check if the literal is valid.
+	if (literal && literal.length) {
+		// Initialize the current position.
+		var currentPosition = 0;
+		// Initialize the variables
+		var index, match, remainingText;
+		// Initialize the variable regular expression.
+		var variable_regex = /(\#|\!)\{(.+?)\}/gi;
+		// Iterate through each match.
+		while ((match = variable_regex.exec(literal)) !== null) {
+			// Delete the input from the match.
+			delete match.input;
+			// Initialize the handlers.
+			var handlers = [];
+			// Check if text is available to be written.
+			if (match.index - currentPosition > 0) {
+				// Emit text.
+				emitter.text(state, literal.substr(currentPosition, match.index - currentPosition), isOpeningElement);
 			}
-			// Increment the index.
-			index += 1;
+			// Check if the variable name contains a function list.
+			if ((index = match[2].search(/[|,]/)) !== -1) {
+				// Split the handler list.
+				var split = match[2].substr(index + 1).split(/[,|]/);
+				// Iterate through each handler.
+				for (var i = 0; i < split.length; i += 1) {
+					// Check if the handler is valid.
+					if (split[i].trim().length) {
+						// Push the handler on the handler list.
+						handlers.push(split[i].trim());
+					}
+				}
+				// Set the variable name.
+				match[2] = match[2].substr(0, index);
+			}
+			// Emit a variable.
+			emitter.variable(state, match[2], handlers, match[1] === '#', isOpeningElement);
+			// Move the current position to the position after the variable.
+			currentPosition = match.index + match[0].length;
+		}
+		// Retrieve the remaining text and check if the remaining text is valid.
+		if ((remainingText = literal.substr(currentPosition)) && remainingText.length) {
+			// Emit text.
+			emitter.text(state, remainingText, isOpeningElement);
 		}
 	}
-	// Return an invalid index.
-	return -1;
+	// Return the state.
+	return state;
+}
+
+/**
+ * Compile the section into invocables and literals.
+ *
+ * @param state The state.
+ * @param section The section.
+ * @param analysis The lexical analysis.
+ * @param position The position of the section.
+ * @returns The state.
+ */
+function compileSection(state, section, analysis, position) {
+	// Check if the section is valid.
+	if (section && section.length) {
+		// Initialize the current position.
+		var currentPosition = position;
+		// Initialize the dispose boolean.
+		var disposed = false;
+		// Iterate through each beginning element.
+		for (var i = 0; i < analysis.beginnings.length; i += 1) {
+			// Retrieve the beginning element.
+			var beginning = analysis.beginnings[i];
+			// Check if the beginning element is an invocation and is contained within the section.
+			if (beginning[3] && beginning.index >= position && beginning.index < position + section.length) {
+				// Compile the literal into variables and text.
+				compileLiteral(state, section.substr(currentPosition - position, beginning.index - currentPosition));
+				// Emit an invocation statement and check if it is not be disposed of.
+				if (!(disposed = emitter.invoke(state, beginning.attributes) && beginning[1] === 'ins')) {
+					// Check if the opening element has attributes.
+					if (beginning.attributes.length) {
+						// Initialize the attributes.
+						var attributes = [];
+						// Iterate through each attribute.
+						for (var j = 0; j < beginning.attributes.length; j += 1) {
+							// Check if the attribute is valid.
+							if (beginning.attributes[j]) {
+								// Combine the key and value for the attribute.
+								attributes.push(beginning.attributes[j].key + '="' + beginning.attributes[j].value + '"');
+							}
+						}
+						// Compile the literal into variables and text.
+						compileLiteral(state, '<' + beginning[1] + (attributes.length === 0 ? '' : ' ' + attributes.join(' ')) + '>');
+					}
+					// Otherwise the opening element has no attributes.
+					else {
+						// Compile the literal into variables and text.
+						compileLiteral(state, '<' + beginning[1] + '>');
+					}
+				}
+				// Move the current position to the position after the invocable.
+				currentPosition = beginning.index + beginning[0].length;
+			}
+		}
+		// Compile the literal into variables and text.
+		compileLiteral(state, section.substr(currentPosition - position));
+	}
+	// Return the state.
+	return state;
+}
+
+/**
+ * Compile the template into associations and sections.
+ *
+ * @param state The state.
+ * @param template The template.
+ * @param analysis The lexical analysis.
+ * @param left The left boundary.
+ * @param right The right boundary.
+ * @returns The state.
+ */
+function compileTemplate(state, template, analysis, left, right) {
+	// Check if the template is valid.
+	if (template && template.length) {
+		// Initialize the current position.
+		var currentPosition = left;
+		// Initialize the dispose boolean.
+		var disposed = false;
+		// Iterate through each association.
+		for (var i = 0; i < analysis.associations.length; i += 1) {
+			// Check if the association is contained within the boundaries.
+			if (analysis.associations[i].beginning.index >= currentPosition && analysis.associations[i].beginning.index >= left && analysis.associations[i].ending.index + analysis.associations[i].ending[0].length <= right) {
+				// Clone the attributes. This allows the writer to use possibly deleted attribute information on element endings.
+				var attributesSliced = analysis.associations[i].beginning.attributes.slice(0);
+				// Compile the section into invocables and literals.
+				compileSection(state, template.substr(currentPosition, analysis.associations[i].beginning.index - currentPosition), analysis, currentPosition);
+				// Emit a beginning statement and check if the element is not be disposed of.
+				if (!(disposed = emitter.begin(state, analysis.associations[i].beginning.attributes) && analysis.associations[i].beginning[1] === 'ins')) {
+					// Check if the opening element has attributes.
+					if (analysis.associations[i].beginning.attributes.length) {
+						// Initialize the attributes.
+						var attributes = [];
+						// Iterate through each attribute.
+						for (var j = 0; j < analysis.associations[i].beginning.attributes.length; j += 1) {
+							// Check if the attribute is valid.
+							if (analysis.associations[i].beginning.attributes[j]) {
+								// Combine the key and value for the attribute.
+								attributes.push(analysis.associations[i].beginning.attributes[j].key + '="' + analysis.associations[i].beginning.attributes[j].value + '"');
+							}
+						}
+						// Process a literal for each variable and remaining text.
+						compileLiteral(state, '<' + analysis.associations[i].beginning[1] + (attributes.length === 0 ? '' : ' ' + attributes.join(' ')) + '>', true);
+					}
+					// Otherwise the opening element has no attributes.
+					else {
+						// Process a literal for each variable and remaining text.
+						compileLiteral(state, '<' + analysis.associations[i].beginning[1] + '>', true);
+					}
+				}
+				// Recursively compile the template into analysis.associations and sections.
+				compileTemplate(state, template, analysis, analysis.associations[i].beginning.index + analysis.associations[i].beginning[0].length, analysis.associations[i].ending.index);
+				// Emit an ending statement.
+				emitter.end(state, attributesSliced);
+				// Move the current position to the position after the association.
+				currentPosition = analysis.associations[i].ending.index + analysis.associations[i].ending[0].length;
+				// Check if the element is not be disposed of and write the ending element.
+				if (!disposed) {
+					// Compile the literal into variables and text.
+					compileLiteral(state, '</' + analysis.associations[i].beginning[1] + '>');
+				}
+			}
+		}
+		// Compile the section into invocables and literals.
+		compileSection(state, template.substr(currentPosition, right - currentPosition), analysis, currentPosition);
+	}
+	// Return the state.
+	return state;
 }
