+++ lib/statements/each-statement.js
@@ -5,21 +5,16 @@
  * @param attribute The attribute.
  */
 module.exports.begin = function (state, attribute) {
-	// Check if the insert position has not been set.
-	if (state.insert === -1) {
-		// Set the insert position.
-		state.insert = state.instructions.length;
-	}
 	// Push the instruction to open a scope for filtering.
-	state.instructions.push('(function (parent) {');
+	state.push('(function (parent) {', 1);
 	// Push the instruction to open the if-statement checking the parent.
-	state.instructions.push('if (parent) {');
+	state.push('if (parent) {', 1);
 	// Push the instruction to open the for-statement.
-	state.instructions.push('for (var key in parent) {');
+	state.push('for (var key in parent) {', 1);
 	// Push the instruction to open the if-statement checking the validity of the property.
-	state.instructions.push('if (parent.hasOwnProperty(key) && parent[key] !== undefined && parent[key] !== null) {');
+	state.push('if (parent.hasOwnProperty(key) && parent[key] !== undefined && parent[key] !== null) {', 1);
 	// Push the instruction to open a scope for the children to run in.
-	state.instructions.push('(function (parent, data) {');
+	state.push('(function (parent, data) {', 1);
 };
 
 /**
@@ -30,9 +25,13 @@ module.exports.begin = function (state, attribute) {
  */
 module.exports.end = function (state, attribute) {
 	// Push the instruction to invoke the scope for the children to run in.
-	state.instructions.push('})(data, parent[key]);');
-	// Push the instruction to close each for- and if-statement.
-	state.instructions.push('}}}');
+	state.push('})(data, parent[key]);', -1);
+	// Push the instruction to close the hasOwnProperty if-statement.
+	state.push('}', -1);
+	// Push the instruction to close the for-statement.
+	state.push('}', -1);
+	// Push the instruction to close the parent if-statement.
+	state.push('}', -1);
 	// Push the instruction to invoke the scope for filtering.
-	state.instructions.push('})(' + attribute.value + ');');
+	state.push('})(' + attribute.value + ');', -1);
 };
