+++ lib/statements/if-statement.js
@@ -5,13 +5,8 @@
  * @param attribute The attribute.
  */
 module.exports.begin = function (state, attribute) {
-	// Check if the insert position has not been set.
-	if (state.insert === -1) {
-		// Set the insert position.
-		state.insert = state.instructions.length;
-	}
 	// Push the instruction to open the if-statement.
-	state.instructions.push('if (' + attribute.value + ') {');
+	state.push('if (' + attribute.value + ') {', 1);
 };
 
 /**
@@ -22,5 +17,5 @@ module.exports.begin = function (state, attribute) {
  */
 module.exports.end = function (state, attribute) {
 	// Push the instruction to close the if-statement.
-	state.instructions.push('}');
+	state.push('}', -1);
 };
