+++ lib/statements/include-statement.js
@@ -10,15 +10,10 @@ var utils = require('../utils');
  * @param attribute The attribute.
  */
 module.exports.begin = function (state, attribute) {
-	// Check if the insert position has not been set.
-	if (state.insert === -1) {
-		// Set the insert position.
-		state.insert = state.instructions.length;
-	}
 	// Increment the depth.
 	state.depth += 1;
 	// Push the instruction to open a scope for output partials.
-	state.instructions.push('(function (outputPartials) {');
+	state.push('(function (outputPartials) {', 1);
 };
 
 /**
@@ -33,9 +28,9 @@ module.exports.end = function (state, attribute) {
 	// Decrement the depth.
 	state.depth -= 1;
 	// Push the instruction to invoke the runtime include.
-	state.instructions.push('result += runtime.render(' + values.data + ', outputPartials, \'' + utils.inline(values.name) + '\');')
+	state.push('result += runtime.render(' + values.data + ', outputPartials, \'' + utils.inline(values.name) + '\');')
 	// Push the instruction to invoke the scope for output partials.
-	state.instructions.push('})([]);');
+	state.push('})([]);', -1);
 };
 
 /**
@@ -48,5 +43,5 @@ module.exports.invoke = function (state, attribute) {
 	// Extract the name- and data values.
 	var values = utils.extract(attribute.value);
 	// Push the instruction to invoke the runtime include.
-	state.instructions.push('result += runtime.render(' + values.data + ', null, \'' + utils.inline(values.name) + '\');')
+	state.push('result += runtime.render(' + values.data + ', null, \'' + utils.inline(values.name) + '\');')
 }
