+++ lib/statements/partial-statement.js
@@ -10,19 +10,14 @@ var utils = require('../utils');
  * @param attribute The attribute.
  */
 module.exports.begin = function (state, attribute) {
-	// Check if the insert position has not been set.
-	if (state.insert === -1) {
-		// Set the insert position.
-		state.insert = state.instructions.length;
-	}
 	// Check if the depth allows for the partial registration.
 	if (state.depth !== 0) {
 		// Push the instruction to open a scope.
-		state.instructions.push('outputPartials[\'' + utils.inline(utils.extract(attribute.value).name) + '\'] = (function (data) {');
+		state.push('outputPartials[\'' + utils.inline(utils.extract(attribute.value).name) + '\'] = (function (data) {', 1);
 		// Push the instruction to open and return the partial function.
-		state.instructions.push('return function () {');
+		state.push('return function () {', 1);
 		// Push the instruction to initialize the result.
-		state.instructions.push('var result = \'\';');
+		state.push('var result = \'\';');
 	}
 	// Decrement the depth.
 	state.depth -= 1;
@@ -40,11 +35,11 @@ module.exports.end = function (state, attribute) {
 	// Check if the depth indicates an inclusion scope.
 	if (state.depth !== 0) {
 		// Push the instruction to return the result.
-		state.instructions.push('return result;');
+		state.push('return result;');
 		// Push the instruction to close the partial function.
-		state.instructions.push('};');
+		state.push('};', -1);
 		// Push the instruction to close and invoke the scope.
-		state.instructions.push('})(' + utils.extract(attribute.value).data + ');');
+		state.push('})(' + utils.extract(attribute.value).data + ');', -1);
 	}
 	// Otherwise this is not in an inclusion scope.
 	else {
@@ -63,5 +58,5 @@ module.exports.invoke = function (state, attribute) {
 	// Extract the name- and data values.
 	var values = utils.extract(attribute.value);
 	// Push the instruction to invoke the runtime partial.
-	state.instructions.push('result += inputPartials && \'' + utils.inline(values.name) + '\' in inputPartials ? inputPartials[\'' + utils.inline(values.name) + '\'](' + values.data + ') : \'\';');
+	state.push('result += inputPartials && \'' + utils.inline(values.name) + '\' in inputPartials ? inputPartials[\'' + utils.inline(values.name) + '\'](' + values.data + ') : \'\';');
 }
