+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "gaikan",
-  "version": "1.3.0",
+  "version": "1.3.1",
   "description": "HTML template engine for Node and Express.",
   "main": "index.js",
   "scripts": {
